pyproject.toml
[project]
name = "dora-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.4.0",
    "fastapi>=0.115.13",
    "gradio>=5.34.2",
    "groq>=0.28.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "opencv-python>=4.11.0.86",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "speechrecognition>=3.14.3",
    "uvicorn>=0.34.3",
]
